+++ tests/testthat/test-tm_t_summary.R
@@ -34,7 +34,7 @@ test_that("template_summary generates correct expressions", {
           na.rm = FALSE,
           na_level = "<Missing>",
           denom = "N_col",
-          .stats = c("n", "mean_sd", "median", "range", "count_fraction")
+          .stats = c("n", "mean_sd", "mean_ci", "median", "median_ci", "range", "count_fraction")
         )
     ),
     table = quote({
@@ -82,7 +82,7 @@ test_that("template_summary can generate customized table", {
           na.rm = TRUE,
           na_level = "<Missing>",
           denom = "N_col",
-          .stats = c("n", "mean_sd", "median", "range", "count")
+          .stats = c("n", "mean_sd", "mean_ci", "median", "median_ci", "range", "count")
         )
     ),
     table = quote({
@@ -134,7 +134,7 @@ test_that("template_summary generates correct expressions for multiple grouping 
           na.rm = FALSE,
           na_level = "<Missing>",
           denom = "N_col",
-          .stats = c("n", "mean_sd", "median", "range", "count_fraction")
+          .stats = c("n", "mean_sd", "mean_ci", "median", "median_ci", "range", "count_fraction")
         )
     ),
     table = quote({
@@ -187,7 +187,60 @@ test_that("template_summary generates correct expressions for multiple grouping 
           na.rm = FALSE,
           na_level = "<Missing>",
           denom = "N_col",
-          .stats = c("n", "mean_sd", "median", "range", "count_fraction")
+          .stats = c("n", "mean_sd", "mean_ci", "median", "median_ci", "range", "count_fraction")
+        )
+    ),
+    table = quote({
+      result <- build_table(lyt = lyt, df = anl, alt_counts_df = adsl)
+      result
+    })
+  )
+  expect_equal(result, expected)
+})
+
+test_that("template_summary generates correct expressions for customized numeric statistics", {
+  result <- template_summary(
+    dataname = "adrs",
+    parentname = "adsl",
+    arm_var = c("ARM", "STRATA1"),
+    sum_vars = c("RACE", "COUNTRY", "AGE"),
+    show_labels = "visible",
+    add_total = FALSE,
+    var_labels = character(),
+    na.rm = FALSE,
+    numeric_stats = c("n"),
+    denominator = "N",
+    drop_arm_levels = TRUE
+  )
+  expected <- list(
+    data = quote({
+      anl <- adrs %>%
+        df_explicit_na(
+          omit_columns = setdiff(names(adrs), c(c("RACE", "COUNTRY", "AGE"))),
+          na_level = "<Missing>"
+        )
+      anl <- anl %>% mutate(ARM = droplevels(ARM))
+      arm_levels <- levels(anl[["ARM"]])
+      adsl <- adsl %>% filter(ARM %in% arm_levels)
+      adsl <- adsl %>% mutate(ARM = droplevels(ARM))
+      anl <- anl %>% mutate(STRATA1 = droplevels(STRATA1))
+      arm_levels <- levels(anl[["STRATA1"]])
+      adsl <- adsl %>% filter(STRATA1 %in% arm_levels)
+      adsl <- adsl %>% mutate(STRATA1 = droplevels(STRATA1))
+      adsl <- df_explicit_na(adsl, na_level = "")
+    }),
+    layout = quote(
+      lyt <- basic_table() %>%
+        split_cols_by("ARM") %>%
+        split_cols_by("STRATA1", split_fun = drop_split_levels) %>%
+        add_colcounts() %>%
+        summarize_vars(
+          vars = c("RACE", "COUNTRY", "AGE"),
+          show_labels = "visible",
+          na.rm = FALSE,
+          na_level = "<Missing>",
+          denom = "N_col",
+          .stats = c("n", "count_fraction")
         )
     ),
     table = quote({