+++ NEWS.md
@@ -17,6 +17,8 @@ variables cannot be overlapped.
 * Added checkbox to `tm_g_pp_patient_timeline` to choose to display relative study days on the x-axis.
 * Limited the label repel feature in `tm_g_pp_patient_timeline` to X-axis for a more consistent look.
 * Added patient ID title to plots in patient profile modules.
+* Added a new argument `numeric_stats` in `tm_t_summary` module so that users can control what statistics to display
+for numeric summarize variables.
 
 # teal.modules.clinical 0.8.8
 ### Enhancements